+++ frontend/src/features/map_planning/components/EditorMap.tsx
@@ -3,8 +3,7 @@ import Konva from 'konva';
 import { useCallback, useContext, useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { ShepherdTourContext } from 'react-shepherd';
-import { toast } from 'react-toastify';
-import { GainedBlossomsDto, LayerDto, LayerType } from '@/api_types/definitions';
+import { LayerDto, LayerType } from '@/api_types/definitions';
 import IconButton from '@/components/Button/IconButton';
 import CancelConfirmationModal from '@/components/Modals/ExtendedModal';
 import {
@@ -23,7 +22,6 @@ import GridIcon from '@/svg/icons/grid-dots.svg?react';
 import RedoIcon from '@/svg/icons/redo.svg?react';
 import TagsIcon from '@/svg/icons/tags.svg?react';
 import UndoIcon from '@/svg/icons/undo.svg?react';
-import { gainBlossom } from '../api/gainBlossom';
 import { useCompleteTour, useReenableTour } from '../hooks/tourHookApi';
 import BaseLayer from '../layers/base/BaseLayer';
 import BaseLayerRightToolbar from '../layers/base/components/BaseLayerRightToolbar';
@@ -167,6 +165,7 @@ export const EditorMap = ({ layers }: MapProps) => {
   }
 
   useEffect(() => {
+    /* Blossoms are removed temporarily and will be added back later
     const _tourCompletionBlossom = async () => {
       const blossom: GainedBlossomsDto = {
         blossom: 'graduation_day',
@@ -177,7 +176,7 @@ export const EditorMap = ({ layers }: MapProps) => {
       toast.success(`${t('blossoms:blossom_gained')} ${t('blossoms:types.graduation_day')}`, {
         icon: '\u{1F338}',
       });
-    };
+    };*/
 
     tour?.start();
     if (tour && tour.steps.length > 0) {
@@ -185,7 +184,7 @@ export const EditorMap = ({ layers }: MapProps) => {
         setShow(true);
       });
       tour?.on('complete', () => {
-        _tourCompletionBlossom();
+        // _tourCompletionBlossom(); Blossoms are removed temporarily and will be added back later
         completeTour();
       });
     }